Added a user-friendly way to enter the video location through command-line interfaces by using argparse.
If you would like to process videos, all you have to do is add to the command line an argument with the directory of the videos.Example:

python main.py ../test_video/

The argument of the directory location is not necessary, but if provided the program will find any mp4 videos and will output the processed videos with the roadlines detected. The program also automatically generates help when users give the program invalid arguments.